# openapi_full.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ConsciousAI Journal",
    "description": "AI-powered self-reflection journaling application",
    "version": "2.0.0a1"
  },
  "paths": {
...
    "/api/v1/journals": {
      "post": {
        "tags": [
          "journals"
        ],
        "summary": "Create a new journal entry with AI reflection and safety checks",
        "description": "Create a new journal entry adhering to strict safety-first order.\n\nSafety Boundary:\n- If input contains crisis ideation or self-harm, pipeline halts immediately.\n- Zero persistence occurs; entry is NOT saved to the database.\n- Safe crisis envelope and regional helplines are returned immediately.\n- If safe, entry is persisted along with AI reflection and classifications.",
        "operationId": "create_journal_entry_api_v1_journals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List user's journal entries with filtering and deterministic pagination",
        "description": "Retrieve paginated journal entries for the current user.",
        "operationId": "list_journal_entries_api_v1_journals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset number of records",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset number of records"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size limit"
          },
          {
            "name": "emotion",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by top emotion",
              "title": "Emotion"
            },
            "description": "Filter by top emotion"
          },
          {
            "name": "value_theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by top value theme",
              "title": "Value Theme"
            },
            "description": "Filter by top value theme"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "is_private",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by privacy visibility",
              "title": "Is Private"
            },
            "description": "Filter by privacy visibility"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start datetime",
              "title": "Start Date"
            },
            "description": "Inclusive start datetime"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end datetime",
              "title": "End Date"
            },
            "description": "Inclusive end datetime"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring search query",
              "title": "Search"
            },
            "description": "Substring search query"
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journals/export": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Export user's journal entries in CSV or JSON format",
        "description": "Export the authenticated user's non-deleted journal entries.",
        "operationId": "export_journal_entries_api_v1_journals_export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/JournalExportFormat",
              "description": "Target export format (csv or json).",
              "default": "json"
            },
            "description": "Target export format (csv or json)."
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start datetime",
              "title": "Start Date"
            },
            "description": "Inclusive start datetime"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end datetime",
              "title": "End Date"
            },
            "description": "Inclusive end datetime"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journals/{entry_id}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get a single journal entry by ID",
        "description": "Fetch a single journal entry owned by the current user.",
        "operationId": "get_journal_entry_api_v1_journals__entry_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "journals"
        ],
        "summary": "Update mutable fields of a journal entry",
        "description": "Update mutable attributes of an existing entry owned by current user.",
        "operationId": "update_journal_entry_api_v1_journals__entry_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journals"
        ],
        "summary": "Delete a journal entry (soft delete by default, permanent optional)",
        "description": "Delete a journal entry owned by the current user.",
        "operationId": "delete_journal_entry_api_v1_journals__entry_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          },
          {
            "name": "permanent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to permanently remove record",
              "default": false,
              "title": "Permanent"
            },
            "description": "Whether to permanently remove record"
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Development-only tenant identifier (ignored in production mode).",
              "title": "X-User-Id"
            },
            "description": "Development-only tenant identifier (ignored in production mode)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Journal Entry Api V1 Journals  Entry Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JournalEntryCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "description": "Raw journal reflection text."
          },
          "mood_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood Score",
            "description": "Optional subjective mood score between 1.0 and 10.0."
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Optional thematic or categorical tags."
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "description": "Companion persona style ('Supportive', 'Coach', 'Therapist-Style Reflection', 'Neutral').",
            "default": "Supportive"
          },
          "region": {
            "type": "string",
            "title": "Region",
            "description": "User region for crisis resource routing ('US', 'CA', 'GB', 'AU', 'IN', 'GLOBAL').",
            "default": "GLOBAL"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private",
            "description": "Privacy visibility flag for the entry.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "JournalEntryCreate",
        "description": "Payload for creating a new journal entry."
      },
      "JournalEntryCreateResponse": {
        "properties": {
          "entry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JournalEntryRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "The persisted journal entry record, or None if halted by safety boundaries."
          },
          "reflection": {
            "$ref": "#/components/schemas/ReflectionResult",
            "description": "Synthesized AI reflection or safe crisis response envelope."
          },
          "input_safety": {
            "$ref": "#/components/schemas/SafetyCheckResult",
            "description": "Input safety classification and crisis resource routing metadata."
          },
          "output_safety": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputSafetyCheckResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Output safety validation results."
          }
        },
        "type": "object",
        "required": [
          "reflection",
          "input_safety"
        ],
        "title": "JournalEntryCreateResponse",
        "description": "Response returned when a journal entry is submitted.\n\nFor safe input: includes the persisted entry and reflection results.\nFor unsafe input: entry is None (zero persistence) and crisis guidance is returned."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}